优化相机拍照获取分辨率问题，增加触摸对焦功能

+++ app/src/main/java/com/example/team/carmeraview/util/DisplayUtil.java
@@ -59,11 +59,11 @@ public class DisplayUtil {
      * @param context
      * @return
      */
-    public static float getScreenHeight(Context context) {
+    public static int getScreenHeight(Context context) {
         return getDisplayMetrics(context).heightPixels;
     }
 
-    public static float getScreenWidth(Context context) {
+    public static int getScreenWidth(Context context) {
         return getDisplayMetrics(context).widthPixels;
     }
